rename EkspanderbarSammenligning -> Sammenligningspaneler, BransjeSammenligningspanel -> Sammenligningspanel
