+++ src/Forside/Sammenligningspaneler/GetBransjeEllerNæringKategori.ts
